Fix format of fntl request

+++ pgxn/neon/pagestore_client.h
@@ -102,7 +102,7 @@ typedef struct
 {
 	NeonRequest req;
 	int cmd;
-	int arg;
+	uint64 arg;
 	int size;
 	char data[FLEXIBLE_ARRAY_MEMBER];
 }			NeonFcntlRequest;
